move initial loading to separate file
